+++ controllers/quickworkout.js
@@ -151,18 +151,30 @@ const deleteQuickworkout = async (req, res) => {
 	}
 
 	try {
-		const deletedQuickworkout = await yogaworkoutQuickworkout.deleteOne({
-			_id: quickworkoutId,
-		});
+		const documentExists = await yogaworkoutQuickworkout.findOne({ _id: quickworkoutId, });
+		if (documentExists) {
+			const deletedQuickworkout = await yogaworkoutQuickworkout.deleteOne({
+				_id: quickworkoutId,
+			});
 
-		if (deletedQuickworkout.deletedCount === 0) {
-			return res.status(404).json({ error: 'Quickworkout not found' });
+			if (deletedQuickworkout.deletedCount === 0) {
+				return res.status(404).json({ error: 'Quickworkout not found' });
+			}
+			else {
+				ImageToDelet = documentExists.image;
+				const imageRes = await deleteFile(ImageToDelet);
+				// console.log("imageRes", imageRes)
+			}
+
+			res.json({
+				message: 'Quickworkout deleted successfully',
+				deletedQuickworkout,
+			});
+		}
+		else {
+			console.log('No document found to delete.');
 		}
 
-		res.json({
-			message: 'Quickworkout deleted successfully',
-			deletedQuickworkout,
-		});
 	} catch (err) {
 		console.error(err);
 		res.status(500).json({ error: 'Failed to delete Quickworkout' });
@@ -183,15 +195,13 @@ const changeQuickworkoutStatus = async (req, res) => {
 	let quickworkoutId = req.body.id.toString();
 	let quickworkoutStatus = req.body.status;
 
-	// console.log('req.body', req.body);
-
 	if (mongoose.Types.ObjectId.isValid(quickworkoutId)) {
 		const updatedQuickworkout = await yogaworkoutQuickworkout.findOneAndUpdate(
 			{ _id: quickworkoutId },
 			{ $set: { isActive: quickworkoutStatus } },
 			{ returnDocument: 'after' }
 		);
-		// console.log('updatedQuickworkout', updatedQuickworkout);
+
 		if (!updatedQuickworkout) {
 			return res.status(404).json({ error: 'Quickworkout not found' });
 		}
